last

+++ ParkV4.Application/Brands/Commands/Delete/DeleteBrandCommand.cs
@@ -40,6 +40,11 @@ public class DeleteBrandCommand : IRequest<BaseResponseModel<Unit>>
             List<Entry> entries = await _context.Entries
                 .Where(c => vehicleIds.Contains(c.VehicleId))
                 .ToListAsync(cancellationToken);
+
+            if (entries.Count > 0)
+            {
+                throw new Exception("Bu markaya ait giriş/çıkış verileri bulunmaktadır. İlgili veriler silinmeden marka silinemez.");
+            }
             
             _context.Entries.RemoveRange(entries);
             await _context.SaveChangesAsync(cancellationToken);
